requirements.txt

Django==5.2.3
djangorestframework==3.15.1
python-dotenv==1.0.1
langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.1
tiktoken==0.7.0
openai==1.30.1
faiss-cpu==1.8.0
PyPDF2==3.0.1
unstructured==0.13.4
markdown==3.6
beautifulsoup4==4.12.3
html2text==2024.2.26
django-extensions==3.2.3
